requirements.txt
milvus[client]
pillow
torch
torchvision 
transformers
fugashi[unidic-lite]